refactor: server start-up & separate test folder
- combine `createServer` and `startServer` functions
- remove unnecessary `start-server.js` file
- simplify jest config
- remove `babel` dependency & config
- remove unused `.npmignore` file (this is not a library)
- separate the tests out into their own directory
