--- jest.config.js
@@ -1,22 +1,0 @@
-export default {
-  rootDir: '.',
-  verbose: true,
-  resetModules: true,
-  clearMocks: true,
-  silent: false,
-  watchPathIgnorePatterns: ['globalConfig'],
-  testMatch: ['**/src/**/*.test.js'],
-  reporters: ['default', ['github-actions', { silent: false }], 'summary'],
-  setupFiles: ['<rootDir>/.jest/setup-files.js'],
-  collectCoverageFrom: ['src/**/*.js'],
-  coveragePathIgnorePatterns: ['<rootDir>/node_modules/', '<rootDir>/.server', 'index.js'],
-  coverageDirectory: '<rootDir>/coverage',
-  transform: {
-    '^.+\\.js$': 'babel-jest'
-  },
-  transformIgnorePatterns: [
-    `node_modules/(?!${[
-      '@defra/hapi-tracing' // Supports ESM only
-    ].join('|')}/)`
-  ]
-}
