+++ jest.config.json
@@ -1,0 +1,10 @@
+{
+  "testMatch": ["<rootDir>/test/**/*.test.js"],
+  "transform": {},
+  "collectCoverage": true,
+  "collectCoverageFrom": ["<rootDir>/app/**/*.js"],
+  "coverageDirectory": "coverage",
+  "coverageReporters": ["text-summary", "lcov"],
+  "verbose": true,
+  "resetMocks": true
+}
