--- src/common/helpers/start-server.test.js
@@ -1,79 +1,0 @@
-import hapi from '@hapi/hapi'
-
-const mockLoggerInfo = jest.fn()
-const mockLoggerError = jest.fn()
-
-const mockHapiLoggerInfo = jest.fn()
-const mockHapiLoggerError = jest.fn()
-
-jest.mock('hapi-pino', () => ({
-  register: (server) => {
-    server.decorate('server', 'logger', {
-      info: mockHapiLoggerInfo,
-      error: mockHapiLoggerError
-    })
-  },
-  name: 'mock-hapi-pino'
-}))
-jest.mock('./logging/logger.js', () => ({
-  createLogger: () => ({
-    info: (...args) => mockLoggerInfo(...args),
-    error: (...args) => mockLoggerError(...args)
-  })
-}))
-
-describe('#startServer', () => {
-  const PROCESS_ENV = process.env
-  let createServerSpy
-  let hapiServerSpy
-  let startServerImport
-  let createServerImport
-
-  beforeAll(async () => {
-    process.env = { ...PROCESS_ENV }
-    process.env.PORT = '3098' // Set to obscure port to avoid conflicts
-
-    createServerImport = await import('../../server.js')
-    startServerImport = await import('./start-server.js')
-
-    createServerSpy = jest.spyOn(createServerImport, 'createServer')
-    hapiServerSpy = jest.spyOn(hapi, 'server')
-  })
-
-  afterAll(() => {
-    process.env = PROCESS_ENV
-  })
-
-  describe('When server starts', () => {
-    let server
-
-    afterAll(async () => {
-      await server.stop({ timeout: 0 })
-    })
-
-    test('Should start up server as expected', async () => {
-      server = await startServerImport.startServer()
-
-      expect(createServerSpy).toHaveBeenCalled()
-      expect(hapiServerSpy).toHaveBeenCalled()
-      expect(mockHapiLoggerInfo).toHaveBeenNthCalledWith(1, 'Server started successfully')
-      expect(mockHapiLoggerInfo).toHaveBeenNthCalledWith(
-        2,
-        'Access your backend on http://localhost:3098'
-      )
-    })
-  })
-
-  describe('When server start fails', () => {
-    beforeAll(() => {
-      createServerSpy.mockRejectedValue(new Error('Server failed to start'))
-    })
-
-    test('Should log failed startup message', async () => {
-      await startServerImport.startServer()
-
-      expect(mockLoggerInfo).toHaveBeenCalledWith('Server failed to start :(')
-      expect(mockLoggerError).toHaveBeenCalledWith(Error('Server failed to start'))
-    })
-  })
-})
